add portal product module

+++ src/main/java/com/mmall/service/impl/FileServiceImpl.java
@@ -1,8 +1,10 @@
 package com.mmall.service.impl;
 
 import com.google.common.collect.Lists;
+import com.mmall.common.ResponseCode;
 import com.mmall.service.FileService;
 import com.mmall.util.FTPUtil;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
@@ -23,12 +25,12 @@ public class FileServiceImpl implements FileService {
     /**
      * 上传文件
      *
-     * @param file
-     * @param path
-     * @return
+     * @param file 上传的文件名
+     * @param path 上传文件的路径
+     * @return 上传成功返回文件名，失败返回error
      */
     @Override
-    public String upload(MultipartFile file, String path) {
+    public String upload(MultipartFile file, String path, String remotePath) {
         //getName获取表单中文件组件的名字
         //getOriginalFilename获取上传文件的原名
         logger.info(file.getOriginalFilename());
@@ -36,8 +38,8 @@ public class FileServiceImpl implements FileService {
         //获取扩展名
         String fileExtensionName = fileName.substring(fileName.lastIndexOf(".") + 1);
         //为上传的文件生成新的文件名
-        String uploadFileName = UUID.randomUUID().toString()+"."+fileExtensionName;
-        logger.info("------开始上传文件,上传的文件名:{},上传的路径:{},新文件名:{}",fileName,path,uploadFileName);
+        String uploadFileName = UUID.randomUUID().toString() + "." + fileExtensionName;
+        logger.info("------开始上传文件,上传的文件名:{},上传的路径:{},新文件名:{}", fileName, path, uploadFileName);
 
         File fileDir = new File(path);
         //如果不存在upload就创建
@@ -49,13 +51,23 @@ public class FileServiceImpl implements FileService {
         File targetFile = new File(path, uploadFileName);
         try {
             file.transferTo(targetFile);
+            if (StringUtils.isBlank(remotePath)) {
+                remotePath = "img";
+//                remotePath = "images";
+            }
             //将targetFile上传到FTP服务器
-            FTPUtil.uploadFile(Lists.newArrayList(targetFile),"images");
+            boolean result = FTPUtil.uploadFile(Lists.newArrayList(targetFile), remotePath);
+            if (!result) {
+                //上传失败
+                logger.error("上传文件失败");
+                return ResponseCode.ERROR.getDesc();
+            }
             //上传完之后，删除upload中的文件
             targetFile.delete();
+            return targetFile.getName();
         } catch (IOException e) {
             logger.error("上传文件异常", e);
         }
-        return targetFile.getName();
+        return ResponseCode.ERROR.getDesc();
     }
 }
